requirements-web.txt
Flask>=2.3.0
requests>=2.31.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
MLB-StatsAPI>=1.5.0
psycopg2-binary>=2.9.0
gunicorn>=21.2.0

